规格参数增删改查

+++ mingrui-shop-service/mingrui-shop-service-xxx/src/main/java/com/ali/shop/service/impl/SpecificationServiceImpl.java
@@ -3,8 +3,11 @@ package com.ali.shop.service.impl;
 import com.ali.shop.base.BaseApiService;
 import com.ali.shop.base.Result;
 import com.ali.shop.dto.SpecGroupDTO;
+import com.ali.shop.dto.SpecParamDTO;
 import com.ali.shop.entity.SpecGroupEntity;
+import com.ali.shop.entity.SpecParamEntity;
 import com.ali.shop.mapper.SpecGroupMapper;
+import com.ali.shop.mapper.SpecParamMapper;
 import com.ali.shop.service.SpecificationService;
 import com.ali.shop.utils.ALiBeanUtil;
 import com.alibaba.fastjson.JSONObject;
@@ -28,6 +31,9 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
 
     @Resource
     private SpecGroupMapper specGroupMapper;
+
+    @Resource
+    private SpecParamMapper specParamMapper;
 
     @Transactional
     @Override
@@ -56,7 +62,48 @@ public class SpecificationServiceImpl extends BaseApiService implements Specific
     @Transactional
     @Override
     public Result<JSONObject> deleteSpecGroupById(Integer id) {
+        //删除规格组之前需要先判断一下当前规格组下是否有规格参数 true不能被删除
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",id);
+        List<SpecParamEntity>specParamEntities=specParamMapper.selectByExample(example);
+        if (specParamEntities.size()>0){
+            return this.setResultSuccess("该规格组有数据无法删除");
+        }
         specGroupMapper.deleteByPrimaryKey(id);
+        return this.setResultSuccess();
+    }
+
+    @Override
+    @Transactional
+    public Result<List<SpecParamEntity>> getSpecParamInfo(SpecParamDTO specParamDTO) {
+        SpecParamEntity specParamEntity = ALiBeanUtil.copyProperties(specParamDTO, SpecParamEntity.class);
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",specParamEntity.getGroupId());
+
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+
+        return this.setResultSuccess(specParamEntities);
+    }
+
+    @Transactional
+    @Override
+    public Result<JSONObject> saveSpecParam(SpecParamDTO specParamDTO) {
+        specParamMapper.insertSelective(ALiBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+
+        return this.setResultSuccess();
+    }
+
+    @Transactional
+    @Override
+    public Result<JSONObject> editSpecParam(SpecParamDTO specParamDTO) {
+        specParamMapper.updateByPrimaryKeySelective(ALiBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Transactional
+    @Override
+    public Result<JSONObject> deleteSpecParam(Integer id) {
+        specParamMapper.deleteByPrimaryKey(id);
         return this.setResultSuccess();
     }
 }
